Revert "1132 naudotoju procesas administravimas naudotoju saraso formos pakeitimai"
