+++ vitrina/users/admin.py
@@ -1,63 +1,30 @@
-import csv
 import secrets
 from datetime import datetime
 
-import pytz
 from allauth.account.models import EmailAddress, EmailConfirmation, EmailConfirmationHMAC
 from allauth.utils import build_absolute_uri
 from django.contrib import admin
 from django.contrib.auth.admin import UserAdmin as BaseUserAdmin
-from django.contrib.contenttypes.models import ContentType
 from django.contrib.sites.models import Site
-from django.db.models import Value, Case, When, CharField, Func, F, Count, Q, Subquery, OuterRef
-from django.db.models.functions import Concat
-from django.http import StreamingHttpResponse
 from django.urls import reverse
 from django.utils.html import format_html
 from django.utils.text import Truncator
 from django.utils.translation import gettext_lazy as _
 
-from vitrina import settings
-from vitrina.filters import FormatFilter
 from vitrina.helpers import email
-from vitrina.orgs.models import Organization
-from vitrina.structure.services import to_row
 from vitrina.users.forms import RegisterAdminForm
 from vitrina.users.models import User
 
 
-class AtTimeZone(Func):
-    function = "AT TIME ZONE"
-    template = f"(%(expressions)s %(function)s '{settings.TIME_ZONE}')"
-
-
 class UserAdmin(BaseUserAdmin):
-    list_display = (
-        'created_display',
-        'last_login_display',
-        'organization_display',
-        'name_display',
-        'email',
-        'status_display',
-    )
-    search_fields = [
-        'first_name',
-        'last_name',
-        'email',
-        'created_formatted',
-        'last_login_formatted',
-        'main_organization',
-        'status_title'
-    ]
-    ordering = ('first_name', 'last_name',)
+    list_display = ('name', 'org', 'last_login', 'is_staff')
+    search_fields = ['first_name', 'last_name', 'email']
+    list_filter = ('is_staff', 'is_superuser', 'is_active', 'organization')
+    ordering = ('email',)
     delete_confirmation_template = "vitrina/users/admin/delete_confirmation.html"
     delete_selected_confirmation_template = "vitrina/users/admin/delete_selected_confirmation.html"
-    change_list_template = 'vitrina/users/admin/change_list.html'
     add_form = RegisterAdminForm
     add_form_template = "vitrina/users/admin/add_form.html"
-    list_display_links = ('name_display',)
-    actions = None
-    list_filter = (FormatFilter,)
 
     fieldsets = (
         (None, {'fields': ('email', 'password')}),
@@ -74,110 +41,20 @@ class UserAdmin(BaseUserAdmin):
         }),
     )
 
-    def get_queryset(self, request):
-        queryset = super().get_queryset(request)
-        queryset = queryset.annotate(full_name=Concat('first_name', Value(' '), 'last_name'))
-        queryset = queryset.annotate(created_with_tz=AtTimeZone(F('created')))
-        queryset = queryset.annotate(created_formatted=Func(
-            F('created_with_tz'),
-            Value('YYYY-MM-DD HH24:MI'),
-            function='to_char',
-            output_field=CharField()
-        ))
-        queryset = queryset.annotate(last_login_with_tz=AtTimeZone(F('last_login')))
-        queryset = queryset.annotate(last_login_formatted=Func(
-            F('last_login_with_tz'),
-            Value('YYYY-MM-DD HH24:MI'),
-            function='to_char',
-            output_field=CharField()
-        ))
-        queryset = queryset.annotate(status_title=Case(
-            When(status=User.ACTIVE, then=Value("Aktyvus")),
-            When(status=User.AWAITING_CONFIRMATION, then=Value("Laukiama patvirtinimo")),
-            When(status=User.SUSPENDED, then=Value("Suspenduotas")),
-            When(status=User.DELETED, then=Value("Pašalintas")),
-            output_field=CharField(),
-        ))
-        queryset = queryset.annotate(organization_rep_count=Count(
-            'representative__pk',
-            filter=Q(representative__content_type=ContentType.objects.get_for_model(Organization))
-        ))
-        queryset = queryset.annotate(main_organization=Case(
-            When(organization_rep_count=1, then=Subquery(
-                Organization.objects.filter(representatives__user_id=OuterRef('pk')).values('title')
-            )),
-            default=F('organization__title')
-        ))
-        return queryset
-
-    def name_display(self, obj: User) -> str:
+    @admin.display()
+    def name(self, obj: User) -> str:
         return f'{obj.first_name} {obj.last_name}'
 
-    name_display.short_description = _('Vardas ir pavardė')
-    name_display.admin_order_field = 'full_name'
-
-    def organization_display(self, obj):
-        reps = obj.representative_set.filter(
-            content_type=ContentType.objects.get_for_model(Organization)
-        )
-        if len(reps) == 1:
-            return format_html(
-                '<a href="{}" target="_blank">{}</a>',
-                reps[0].content_object.get_absolute_url(),
-                Truncator(reps[0].content_object).chars(42),
-            )
-        elif obj.organization:
+    @admin.display()
+    def org(self, obj):
+        if obj.organization:
             return format_html(
                 '<a href="{}" target="_blank">{}</a>',
                 obj.organization.get_absolute_url(),
                 Truncator(obj.organization).chars(42),
             )
         else:
-            return '-'
-    organization_display.short_description = _("Organizacija")
-    organization_display.admin_order_field = 'main_organization'
-
-    def created_display(self, obj):
-        if obj.created:
-            tz = pytz.timezone(settings.TIME_ZONE)
-            return obj.created.astimezone(tz).strftime("%Y-%m-%d %H:%M")
-        return "-"
-    created_display.short_description = _("Sukurtas")
-    created_display.admin_order_field = 'created'
-
-    def last_login_display(self, obj):
-        if obj.last_login:
-            tz = pytz.timezone(settings.TIME_ZONE)
-            return obj.last_login.astimezone(tz).strftime("%Y-%m-%d %H:%M")
-        return "-"
-    last_login_display.short_description = _("Paskutinis prisijungimas")
-    last_login_display.admin_order_field = 'last_login'
-
-    def status_display(self, obj):
-        if obj.status:
-            if obj.status == User.ACTIVE:
-                return format_html(
-                    '<span style="color: lightgreen;">{}</span>',
-                    obj.get_status_display(),
-                )
-            elif obj.status == User.AWAITING_CONFIRMATION:
-                return format_html(
-                    '<span style="color: orange;">{}</span>',
-                    obj.get_status_display(),
-                )
-            elif obj.status == User.SUSPENDED:
-                return format_html(
-                    '<span style="color: red;">{}</span>',
-                    obj.get_status_display(),
-                )
-            elif obj.status == User.DELETED:
-                return format_html(
-                    '<span style="color: red;">{}</span>',
-                    obj.get_status_display(),
-                )
-        return "-"
-    status_display.short_description = _("Būsena")
-    status_display.admin_order_field = 'status'
+            return ''
 
     def add_view(self, request, form_url='', extra_context=None):
         if extra_context is None:
@@ -210,62 +87,6 @@ class UserAdmin(BaseUserAdmin):
         )
 
         return super().response_add(request, obj, post_url_continue)
-
-    def changelist_view(self, request, extra_context=None):
-        extra_context = {
-            'title': _("Naudotojų sąrašas")
-        }
-        result = super().changelist_view(request, extra_context)
-        if request.GET.get('format') and request.GET.get('format') == 'csv':
-            if change_list := result.context_data.get('cl'):
-                stream = self._export_user_list(change_list.queryset)
-                result = StreamingHttpResponse(stream, content_type='text/csv')
-                result['Content-Disposition'] = 'attachment; filename=Naudotojai.csv'
-        elif result.context_data.get('cl'):
-            result.context_data['cl'].has_filters = False
-        return result
-
-    def _export_user_list(self, queryset):
-        class _Stream(object):
-            def write(self, value):
-                return value
-
-        cols = {
-            'created': _("Sukurtas"),
-            'last_login': _("Paskutinis prisijungimas"),
-            'organization': _("Organizacija"),
-            'full_name': _("Vardas ir pavardė"),
-            'email': _("Elektroninis paštas"),
-            'status': _("Būsenas"),
-        }
-        rows = self._get_user(cols, queryset)
-        rows = ({v: row[k] for k, v in cols.items()} for row in rows)
-
-        stream = _Stream()
-        yield stream.write(b'\xef\xbb\xbf')
-
-        writer = csv.DictWriter(stream, fieldnames=cols.values(), delimiter=';')
-        yield writer.writeheader()
-        for row in rows:
-            yield writer.writerow(row)
-
-    def _get_user(self, cols, queryset):
-        for item in queryset:
-            reps = item.representative_set.filter(
-                content_type=ContentType.objects.get_for_model(Organization)
-            )
-            if len(reps) == 1:
-                organization = reps[0].content_object
-            else:
-                organization = item.organization
-            yield to_row(cols.keys(), {
-                'created': self.created_display(item),
-                'last_login': self.last_login_display(item),
-                'organization': organization.title if organization else "-",
-                'full_name': self.name_display(item),
-                'email': item.email,
-                'status': item.get_status_display() if item.status else "-",
-            })
 
 
 admin.site.register(User, UserAdmin)
