+++ vitrina/users/apps.py
@@ -5,4 +5,4 @@ from django.utils.translation import gettext_lazy as _
 class UsersConfig(AppConfig):
     name = 'vitrina.users'
     label = 'vitrina_users'
-    verbose_name = _("Naudotojai")
+    verbose_name = _("Visi naudotojai")
