--- vitrina/users/templates/vitrina/users/admin/change_list.html
@@ -1,22 +1,0 @@
-{% extends "admin/change_list.html" %}
-{% load i18n %}
-
-{% block object-tools-items %}
-<li><a class="addlink" href="{% url 'admin:vitrina_users_user_add' %}">
-    {% translate "Sukurti naują naudotoją" %}
-</a></li>
-<li>
-<form id="export"action="{% url 'admin:vitrina_users_user_changelist' %}" method="get" style="padding: 0px; margin: 0px;">
-    <input type="hidden" name="format" value="csv"/>
-    {% for k, v in request.GET.items %}
-        {% if k != 'format' %}
-            <input type="hidden" name="{{ k }}" value="{{ v | escape }}"/>
-        {% endif %}
-    {% endfor %}
-    <button type="submit" title="{% translate 'Eksportuoti' %}" aria-label="{% translate 'Eksportuoti' %}"
-            style="background-color: #00bbff !important; color: white !important; padding: 10px 20px !important; float: right;">
-        {% translate "Eksportuoti" %}
-    </button>
-</form>
-</li>
-{% endblock %}
